refactor: split and DRY code
DRY-up the callback handler code.
Move CallbackHandlerBuilder to a dedicated file
