+++ src/callback_handler/oci.rs
@@ -1,4 +1,6 @@
 use anyhow::{anyhow, Result};
+use cached::proc_macro::cached;
+use kubewarden_policy_sdk::host_capabilities::oci::ManifestDigestResponse;
 use policy_fetcher::oci_distribution::Reference;
 use policy_fetcher::{registry::Registry, sources::Sources};
 
@@ -28,4 +30,32 @@ impl Client {
         serde_json::to_string(&image_digest)
             .map_err(|e| anyhow!("Cannot serialize response to json: {}", e))
     }
+}
+
+// Interacting with a remote OCI registry is time expensive, this can cause a massive slow down
+// of policy evaluations, especially inside of PolicyServer.
+// Because of that we will keep a cache of the digests results.
+//
+// Details about this cache:
+//   * only the image "url" is used as key. oci::Client is not hashable, plus
+//     the client is always the same
+//   * the cache is time bound: cached values are purged after 60 seconds
+//   * only successful results are cached
+#[cached(
+    time = 60,
+    result = true,
+    sync_writes = true,
+    key = "String",
+    convert = r#"{ format!("{}", img) }"#,
+    with_cached_flag = true
+)]
+pub(crate) async fn get_oci_digest_cached(
+    oci_client: &Client,
+    img: &str,
+) -> Result<cached::Return<ManifestDigestResponse>> {
+    oci_client
+        .digest(img)
+        .await
+        .map(|digest| ManifestDigestResponse { digest })
+        .map(cached::Return::new)
 }
